requirements.txt
fastapi
uvicorn
playwright
langchain
langchain-openai
langchain-experimental
langchain-community
pandas
requests
beautifulsoup4
pdfplumber
matplotlib
python-multipart
python-dotenv
httpx
tiktoken
openai
pydub
